+++ src/components/AssetsFilters/AssetsFilters.scss
