+++ src/components/AssetsPage/AssetsPage.scss
